home screen and all transaction logic done

+++ lib/model/chart_data.dart
@@ -1,5 +1,5 @@
 class ChartData {
   final String xData;
-  final int yData;
+  final double yData;
   ChartData({required this.xData, required this.yData});
 }
